+++ src/deprecations.jl
@@ -1,7 +1,0 @@
-@deprecate fill fillpath
-@deprecate stroke strokepath
-@deprecate polybbox boundingbox
-@deprecate bboxesintersect boundingboxesintersect
-@deprecate norm distance
-@deprecate dot dotproduct
-@deprecate axes rulers